Start command attributes
Also comes with default attributes.

Basically you can specify a type:
* (i)mpulse
* (c)hain
* (r)epeat

Whether the command block is conditional:
* (!) unconditional
* (?) conditional

And whether the command block is automatic:
* (1) automatic
* (0) needs redstone

It also validates the attributes:
* Errors on duplicate characters (e.g. ii)
* Errors on invalid characters (e.g. k)
* Errors on contradictory characters (e.g. ci or !?)

It's all still in one method, parseAttributes, and not really used in the real
program anywhere, but the code's set up for that, so.. progress!