Fix a bug; change @toLabel to ->toLabel (syntax)
The reason for changing @toLabel to ->toLabel is command selectors of course!
I somehow managed to forget about them. Think @a - does that mean the *label*
a or the *selector* a? Well, the selector, since this is trying to be as close
to command blocks as possible. I had to change the toLabel syntax.

-> hopefully isn't too confusing. It only barely looks different: before, it
was "at label", now it's an arrow pointing to the label/position.

Also, I fixed a small bug about functions eating up the code that came right
after the declaration. No good!