Fix editor tab key bug
When switching tabs via ctrl-tab it would behave as though you pressed the
control key, so ctrl-tab would indent the current line (and switch to the next
tab), and ctrl-shift-tab would de-indent the current line (and switch to the
previous tab). This commit simply prevents tab from having an effect if the
ctrl key is pressed.